Use standard integer types to enhance portability
Integer types in <stdint.h> provides better readability & portability
than "long long".

+++ client.c
@@ -1,4 +1,5 @@
 #include <fcntl.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -9,7 +10,7 @@
 
 int main()
 {
-    long long sz;
+    ssize_t sz;
 
     char buf[1];
     char write_buf[] = "testing writing";
@@ -23,7 +24,7 @@ int main()
 
     for (int i = 0; i <= offset; i++) {
         sz = write(fd, write_buf, strlen(write_buf));
-        printf("Writing to " FIB_DEV ", returned the sequence %lld\n", sz);
+        printf("Writing to " FIB_DEV ", returned the sequence %zd\n", sz);
     }
 
     for (int i = 0; i <= offset; i++) {
@@ -31,7 +32,7 @@ int main()
         sz = read(fd, buf, 1);
         printf("Reading from " FIB_DEV
                " at offset %d, returned the sequence "
-               "%lld.\n",
+               "%zd.\n",
                i, sz);
     }
 
@@ -40,7 +41,7 @@ int main()
         sz = read(fd, buf, 1);
         printf("Reading from " FIB_DEV
                " at offset %d, returned the sequence "
-               "%lld.\n",
+               "%zd.\n",
                i, sz);
     }
